README.md
# Battery Status App

app that shows your battery level written in plain JavaScript

## Run your localserver

```
$ http-server ./
```

## Info

The application shows nice icon and battery level in percents %. If you are plugged in to charge it changes icon.
It is using ionic icons. There are 3 different icons to show - lower than 7% battery - empty icon, 
between 7% and 20% - the low-battery-icon, from 50% up to 80% half-battery-icon, above 80% - full-battery-icon,
and if charging - the charging-battery-icon.